Separated logic into service for authentication
