+++ backend/src/auth/dtos/authenticateUser.servicePayload.ts
@@ -1,0 +1,4 @@
+export type AuthenticateUserServicePayload = {
+  login: string;
+  pin: string;
+};
